move maxFee tests
