+++ test/main/transaction/index.test.js
@@ -240,32 +240,6 @@ describe('#londonToLegacy', () => {
   })
 })
 
-describe('#maxFee', () => {
-  it('sets the max fee as 2 ETH on mainnet', () => {
-    const tx = {
-      chainId: addHexPrefix((1).toString(16))
-    }
-
-    expect(maxFee(tx)).toBe(2e18)
-  })
-
-  it('sets the max fee as 250 FTM on Fantom', () => {
-    const tx = {
-      chainId: addHexPrefix((250).toString(16))
-    }
-
-    expect(maxFee(tx)).toBe(250e18)
-  })
-
-  it('sets the max fee as 50 on other chains', () => {
-    const tx = {
-      chainId: addHexPrefix((255).toString(16))
-    }
-
-    expect(maxFee(tx)).toBe(5e19)
-  })
-})
-
 describe('#populate', () => {
   let gas
   let rawTx
